% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{adapt_mean_stability}
\alias{adapt_mean_stability}
\title{Adapt mean stability}
\usage{
adapt_mean_stability(inference_collection, mean.stability = 0.8, plot = FALSE)
}
\arguments{
\item{inference_collection}{Collection of inferences, result from the all_inferences() function.}

\item{mean.stability}{Value for the mean stability.}

\item{plot}{Boolean.}
}
\value{
A list containing
\itemize{
\item{freqs: }{a tibble gathering the edge selection frequencies corresponding to
the adapted stability of each method.}
\item{stab_data: }{a tibble giving the adapted stability and density of each method.}
}
}
\description{
Adapts each method's stability via the edge density, so that the mean stability is
the desired value.
Adapting the mean stability of the set of methods aims at adapting each method to
the same precision level.
}
\examples{
data("liver")
abund<-liver$abundances
meta=liver$metadata
taxo <- liver$taxonomy
counts_from_table<-get_count_table(abund.table=abund, prev.min=0.9, verbatim=FALSE)$data
some_inferences<-liver$infer_prev90
adapted_stab_data<-adapt_mean_stability(some_inferences, mean.stability=0.8)
}
