% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{all_inferences}
\alias{all_inferences}
\title{Perform network inference with several methods}
\usage{
all_inferences(
  data,
  edge_thresh = 0.9,
  rep.num = 100,
  n.levels = 100,
  Offset = TRUE,
  covar = NULL,
  cores = 1,
  methods = c("PLNnetwork", "SpiecEasi", "gCoda", "EMtree", "Magma", "SPRING", "ZiLN"),
  seed = 10010,
  subsample.ratio = NULL
)
}
\arguments{
\item{data}{Count data set.}

\item{edge_thresh}{Threshold on selection frequencies to set the number of selected edges.}

\item{rep.num}{Number of subsamples.}

\item{n.levels}{Size of probability grid.}

\item{Offset}{Boolean for the computation of offset terms. If TRUE uses GMPR, or RLE on failure.}

\item{covar}{Covariate data set.}

\item{cores}{Number of cores for parallel computation.}

\item{methods}{Vector of characters indicating with which methods the inference should be
performed among "PLNnetwork","SpiecEasi","gCoda","EMtree","Magma", "ZiLN", and "SPRING".}

\item{seed}{A numerical seed for subsampling (required by the SPRING package).}

\item{subsample.ratio}{The subsampling ratio. The default value is 10*sqrt(n)/n when n>144 and 0.8 when n<=144, where n is the sample size.}
}
\value{
A list of outputs from the desired individual methods.
}
\description{
Perform network inference with several methods
}
\examples{
data("liver")
abund<-liver$abundances
meta=liver$metadata
taxo <- liver$taxonomy
counts_from_table<-get_count_table(abund.table=abund, prev.min=0.9, verbatim=FALSE)$data
some_inferences<-all_inferences(counts_from_table, rep.num=4)
str(some_inferences, max.level=1)
}
