% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_CORE}
\alias{clustering_CORE}
\title{Perform clustering of square matrices according to CORE-clustering}
\usage{
clustering_CORE(
  data,
  edge_freq,
  freq.thresh = 0.9,
  selected_taxo,
  min.thresh,
  palette = palette_default()
)
}
\arguments{
\item{data}{Count data set.}

\item{edge_freq}{Edge final frequency, typically the result of an inference aggregation.}

\item{freq.thresh}{Threshold of edge detection.}

\item{selected_taxo}{Taxonomic information about the MGS of the data set.}

\item{min.thresh}{Minimal number of variables per cluster.}

\item{palette}{vector of colours defining the cluster palette.}
}
\value{
A list containing
\itemize{
\item{groups: }{vector of group membership.}
\item{species_groups: }{list of each group's taxonomic information (name of msp, species,
genus, and phylum).}
\item{graph: }{graphical output, the network with nodes coloured according to their group.}
\item{correlation: }{partial correlation matrix}
\item{table_groups: }{list of msp name per group.}
}
}
\description{
Perform clustering of square matrices according to CORE-clustering
}
\examples{
data("liver")
abund<-liver$abundances
meta<-liver$metadata
taxo=liver$taxonomy \%>\% dplyr::select(msp_name, species)
counts_from_table<-get_count_table(abund.table=abund, prev.min=0.9,verbatim=TRUE)$data
species<-liver$msp_set_90
inf_90<-liver$infer_prev90
adapt<-adapt_mean_stability(inf_90, mean.stability=0.8,plot=TRUE)
selected_taxo= dplyr::left_join(data.frame(msp_name=species),taxo, by="msp_name")
 mean<-compute_aggreg_measures(adapt$freqs)$mean
 clustering_output<-clustering_CORE(counts_from_table, mean, freq.thresh=0.9,
 selected_taxo=selected_taxo, min.thresh=3)
}
