% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_aggreg_measures}
\alias{compute_aggreg_measures}
\title{Aggregation}
\usage{
compute_aggreg_measures(data)
}
\arguments{
\item{data}{A table of size (number of edges)x(number of methods), filled with edges selection frequencies.}
}
\value{
A data.frame with the different aggregations in column.
}
\description{
Compute several aggregations: mean, norme 2, Inverse Variance Weighting (IVW), and number of methods detecting the edge (nb.high.freq).
}
\examples{
data("liver")
inf_90<-liver$infer_prev90
adapt<-adapt_mean_stability(inf_90, mean.stability=0.7,plot=TRUE)
aggreg<-compute_aggreg_measures(adapt$freqs)
head(aggreg)
}
