% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimuEval.R
\name{evalQuali}
\alias{evalQuali}
\title{Computes precision and recall values}
\usage{
evalQuali(inf_output, G, edge_thresh = 0.9, aim_stab = 0.9)
}
\arguments{
\item{inf_output}{Complete ouput from the \code{all_inferences()} function.}

\item{G}{Original graph structure to retrieve.}

\item{edge_thresh}{Threshold on selection frequency for edge detection.}

\item{aim_stab}{Stability threshold.}
}
\value{
A tibble summarising performance (precision, recall, stability, number of edges) for all
penalty levels of all individual inferences in the provided \code{inf_output} object.
}
\description{
Computes precision and recall values
}
\examples{
data("liver")
real_data<-liver$abundances
meta=liver$metadata
taxo=liver$taxonomy
count_data<-get_count_table(abund.table=real_data, mgs=taxo$msp_name, prev.min=0.9)$data
simulation<-new_synth_data(count_data, graph_type="cluster",n=50, plot=TRUE)
some_inferences<-all_inferences(simulation$counts , rep.num=4,cores=2)

Eval_inference<-evalQuali(some_inferences,simulation$G)
str(Eval_inference)
Eval_inference\%>\% ggplot(aes(TPR, PPV, color=method))+geom_point()+geom_line()+theme_light()
}
