% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{get_count_table}
\alias{get_count_table}
\title{Count Table}
\usage{
get_count_table(
  abund.path = NULL,
  sample_id = NULL,
  abund.table = NULL,
  prev.min,
  verbatim = TRUE,
  mgs = NULL
)
}
\arguments{
\item{abund.path}{Path to the abundance table in MGP server.}

\item{sample_id}{Ids of samples to keep in the final table.}

\item{abund.table}{Abundance table. By default this table should have the mgs names as first
column, and be in tsv or rds format.}

\item{prev.min}{Double between 0 and 1. Minimal prevalence threshold of MGS to keep in final
table.}

\item{verbatim}{Boolean controlling verbosity.}

\item{mgs}{Character vector giving MGS names, if they are not specified in the abundance table
first column.}
}
\value{
A list containing
\itemize{
\item{data: }{the final count table (tibble).}
\item{prevalences: }{a tibble gathering the prevalence of each MGS.}
}
}
\description{
Get count table from MGS normalized abundances.
}
\details{
Reads data from a provided path to MGP server or the abundance table directly.
Extracts species and transforms abundances into counts by setting the minimal positive
observed value to 1.
}
\examples{
# With table
data("liver")
abund<-liver$abundances
meta=liver$metadata
taxo=liver$taxonomy
count_from_table<-get_count_table(abund.table=abund,  mgs=taxo$msp_name,
prev.min=0.9, verbatim=FALSE)
str(count_from_table)
}
