% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_par_corr}
\alias{get_par_corr}
\title{Wrapper for partial correlations computation}
\usage{
get_par_corr(edge_freq, Sigma, n, freq.thresh = 0.9)
}
\arguments{
\item{edge_freq}{Edge final frequency, typically the result of an inference aggregation.}

\item{Sigma}{Estimation of the empirical sample covariance matrix.}

\item{n}{Number of samples.}

\item{freq.thresh}{Threshold of edge detection.}
}
\value{
Partial correlations between all variables, in vector format.
}
\description{
Wrapper for partial correlations computation
}
\examples{
data("liver")
abund<-liver$abundances
meta<-liver$metadata
taxo=liver$taxonomy
meta=liver$metadata
counts_from_table<-get_count_table(abund.table=abund, prev.min=0.9, verbatim=FALSE)$data
n<-nrow(counts_from_table)

#first estimate Sigma thanks to SPRING model
qdat <- SPRING::mclr(counts_from_table)
Sigma <- mixedCCA::estimateR(qdat, type = "trunc", method = "approx",tol = 1e-6,
 verbose = FALSE)$R

# get the norme 2 aggregation from the inference using all methods, with mean stbaility 0.7
data("liver")
inf_90<-liver$infer_prev90
adapt<-adapt_mean_stability(inf_90, mean.stability=0.8,plot=TRUE)
mean<-compute_aggreg_measures(adapt$freqs)$mean

# compute partial correlations
 ParCor<-get_par_corr(mean,Sigma,n,freq.thresh=0.9)
}
