% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{just_magma}
\alias{just_magma}
\title{Wrapper for Magma}
\usage{
just_magma(
  data,
  rep.num = 100,
  n.levels = 100,
  edge_thresh = 0.9,
  Offset = TRUE,
  covar = NULL,
  subsample.ratio = NULL,
  ...
)
}
\arguments{
\item{data}{Count data set.}

\item{rep.num}{Number of subsamples.}

\item{n.levels}{Size of probability grid.}

\item{edge_thresh}{Threshold on selection frequencies to set the number of selected edges.}

\item{Offset}{Vector of offsets.}

\item{covar}{Covariate data set.}

\item{subsample.ratio}{The subsampling ratio. The default value is 10*sqrt(n)/n when n>144 and 0.8 when n<=144, where n is the sample size.}

\item{...}{Other options used to in the wrappers}
}
\value{
A list containing
\itemize{
\item{output: }{complete output from \code{rMAGMA::magma()}.}
\item{lambda_stab: }{the tibble gathering the stability and number of detected edges per level of penalty.}
}
}
\description{
Wrapper for Magma
}
\examples{
\dontrun{
data("liver")
abund<-liver$abundances
meta=liver$metadata
taxo=liver$taxonomy
counts_from_table<-get_count_table(abund.table=abund, prev.min=0.9, verbatim=FALSE)$data
Magma_output<-just_magma(counts_from_table, rep.num=2)
str(Magma_output, max.level=2)
}
}
