% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{just_spring}
\alias{just_spring}
\title{Wrapper for SPRING}
\usage{
just_spring(
  data,
  rep.num = 100,
  n.levels = 100,
  cores = 1,
  edge_thresh = 0.9,
  covar = NULL,
  seed = 10010,
  subsample.ratio = NULL,
  ...
)
}
\arguments{
\item{data}{Count data set.}

\item{rep.num}{Number of subsamples.}

\item{n.levels}{Size of probability grid.}

\item{cores}{Number of cores for parallel computation.}

\item{edge_thresh}{Threshold on selection frequencies to set the number of selected edges.}

\item{covar}{Covariate data set.}

\item{seed}{A numeric seed for subsampling (native from the original SPRING package).}

\item{subsample.ratio}{The subsampling ratio. The default value is 10*sqrt(n)/n when n>144 and 0.8 when n<=144, where n is the sample size.}

\item{...}{Other options used to in the wrappers}
}
\value{
A list containing
\itemize{
\item{output: }{output value from \code{SPRING::SPRING()}, result from  \code{pulsar::pulsar()} based on StARS criterion.}
\item{lambda_stab: }{the tibble gathering the stability and number of detected edges per level of penalty.}
}
}
\description{
Wrapper for SPRING
}
\examples{
\dontrun{
data("liver")
abund<-liver$abundances
taxo=liver$taxonomy
meta=liver$metadata
counts_from_table<-get_count_table(abund.table=abund, mgs=taxo$msp_name,
prev.min=0.9, verbatim=FALSE)$data
spring_output<-just_spring(counts_from_table, rep.num=2, cores=1)
str(spring_output, max.level=2)
}
}
