% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneNet-package.R
\docType{package}
\name{OneNet-package}
\alias{OneNet}
\alias{OneNet-package}
\title{OneNet: Aggregating network inference methods for the identification of microbial guilds}
\description{
OneNet is a stability based network inference aggregation method from whole metagenome sequencing data, which aims at fostering reproducibility and precision.
}
\author{
\strong{Maintainer}: Raphaëlle Momal \email{raphaelle.momal@inrae.fr}

Authors:
\itemize{
  \item Camille Champion \email{camille.champion@u-paris.fr}
  \item Mathilde Sola \email{mathilde.sola@inrae.fr}
}

Other contributors:
\itemize{
  \item Mahendra Mariadassou \email{mahendra.mariadassou@inrae.fr} (\href{https://orcid.org/0000-0003-2986-354X}{ORCID}) [contributor]
}

}
\keyword{internal}
