% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_SBM}
\alias{clustering_SBM}
\title{Perform clustering of square matrices according to Stochastic Bloc Modelling}
\usage{
clustering_SBM(
  edge_freq,
  freq.thresh = 0.7,
  selected_taxo,
  force.nb.groups = NULL,
  cores = 2,
  as.tibble = FALSE
)
}
\arguments{
\item{edge_freq}{Edge final frequency, typically the result of an inference aggregation.}

\item{freq.thresh}{Threshold of edge detection.}

\item{selected_taxo}{Taxonomic information about the MGS of the data set.}

\item{force.nb.groups}{Boolean to force the number of groups.}

\item{cores}{Cores for parallel computing.}

\item{as.tibble}{Boolean to render the value \code{species_groups} as a tibble.}
}
\value{
A list containing
\itemize{
\item{groups_features: }{number of groups of the best partition found by SBM.}
\item{groups: }{vector of group membership.}
\item{species_groups: }{list (or tibble if as.tibble=TRUE) of each group's taxonomic information (name of msp, species,
genus, and phylum).}
\item{graph: }{graphical output, the network with nodes coloured according to their group.}
}
}
\description{
Perform clustering of square matrices according to Stochastic Bloc Modelling
}
\examples{
data("liver")
abund<-liver$abundances
meta<-liver$metadata
taxo=liver$taxonomy
counts_from_table<-get_count_table(abund.table=abund, mgs=taxo$msp_name,
prev.min=0.9,verbatim=TRUE)$data
species<-liver$msp_set_90

inf_90<-liver$infer_prev90
adapt<-adapt_mean_stability(inf_90, mean.stability=0.7,plot=TRUE)

taxo <- liver$taxonomy
selected_taxo=dplyr::left_join(data.frame(msp_name=species),taxo, by="msp_name")
 norm2<-compute_aggreg_measures(adapt$freqs)$norm2
 clustering_output<-clustering_SBM(norm2, selected_taxo=selected_taxo)
}
